*add speed in the game and fix some stuff

+++ include/Cloud.hh
@@ -15,7 +15,7 @@ private:
 	float _yCloud[10];
 	int _typeCloud[10];
 	int _position;
-	int _speed;
+	float _speed;
 	bool _statut;
 	vita2d_texture* _texture;
 
@@ -26,11 +26,11 @@ public:
 	//Setter
 	void setTexture(vita2d_texture* texture);
 	void setStatut(const bool statut);
-	void setSpeed(const int speed);
+	void setSpeed(const float speed);
 
 	//Getter
 	const bool getStatut() const;
-	const int getSpeed() const;
+	const float getSpeed() const;
 
 	// Display
 	void displayCloud();
